Added range based constructor and kNNsearch
kNN search now performs a search that will search for the most frequently occured points from a given point. 
range based constructor is working now to make the tree better balanced for better performance(closer to logn)
TO DO: a retangular or square based search for returning all the points in that rectangle.